Extract Store::{delete_jobs, purge_job} etc into impl module.
Following on from e2df8f64b0eeaee2b51da507bc780be430b2c727 extracting
more Store methods, this time those focused on deletion, both explicitly
through the API and automatically by the reaper.
