+++ src/store/mod.rs
@@ -4,6 +4,7 @@
 mod complete;
 mod complete_batcher;
 mod cron;
+mod delete;
 mod enqueue_batcher;
 mod fail;
 mod group_committer;
